[#13] remove: UserEntity와 Domain 분리
