+++ TodaysFail-Common/src/main/java/com/todaysfail/common/type/user/AccountRole.java
@@ -1,13 +1,13 @@
-package com.todaysfail.user.domain;
+package com.todaysfail.common.type.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum OauthProvider {
-    KAKAO("KAKAO"),
-    ;
+public enum AccountRole {
+    USER("USER"),
+    ADMIN("ADMIN");
 
     private String value;
 }
