--- TodaysFail-Domain/src/main/java/com/todaysfail/user/domain/OauthInfo.java
@@ -1,18 +1,0 @@
-package com.todaysfail.user.domain;
-
-import javax.persistence.Embeddable;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OauthInfo {
-    @Enumerated(EnumType.STRING)
-    private OauthProvider provider;
-
-    private String oid;
-}
