+++ TodaysFail-Domain/src/main/java/com/todaysfail/user/domain/User.java
@@ -1,31 +1,15 @@
 package com.todaysfail.user.domain;
 
-import javax.persistence.Embedded;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import com.todaysfail.common.type.user.AccountRole;
+import com.todaysfail.common.type.user.AccountStatus;
+import com.todaysfail.common.type.user.OauthProvider;
 
-@Getter
-@Entity(name = "tbl_user")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String nickname;
-
-    @Embedded private OauthInfo oauthInfo;
-
-    @Enumerated(EnumType.STRING)
-    private AccountStatus accountStatus = AccountStatus.NORMAL;
-
-    @Enumerated(EnumType.STRING)
-    private AccountRole accountRole = AccountRole.USER;
+public record User(
+		Long userId,
+		String nickname,
+		OauthProvider oauthProvider,
+		String oid,
+		AccountStatus accountStatus,
+		AccountRole accountRole
+) {
 }
